pyproject.toml
[project]
name = "pc4-eda-kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python>=2.3.0",
    "redis>=7.1.0",
]
